implemented modal on the delete button on edit post page

+++ public/javascripts/post-delete.js
@@ -5,63 +5,41 @@ const deletePostIdFetch = async (postId) => {
     });
 
     if (!res.ok) {
-      console.log("shidddeeddededeededed and fwarteddddddddd");
       throw res;
     }
     const userId = await res.json();
-    const id = {userId}
-    console.log('aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa')
+    const id = { userId }
     window.location.href = `/users/${id}`;
   } catch (err) {
-    console.error(err);
+    const error = await err.json();
+    console.log(error);
   }
 };
 
 document.addEventListener("DOMContentLoaded", (e) => {
-  const deleteModal = document.querySelector('#confirmDeleteModal');
   const pathArr = window.location.pathname.split("/");
   const postId = pathArr[2];
+
+  const containerDiv = document.querySelector('#confirmDeleteModal');
   const deleteBtn = document.querySelector('#delete_post_button');
+  // send delete fetch request to backend and redirect to the user's prof page
   deleteBtn.addEventListener('click', e => {
     e.preventDefault()
-    const confirmDelete = document.getElementById('deletePost')
+    containerDiv.classList.remove('hidden');
+    const okDeleteBtn = document.getElementById('deletePost');
+    const cancelBtn = document.querySelector('#cancel-post-delete');
 
-    confirmDelete.addEventListener('click', (e) => {
-       console.log('bfiujbfszz');
-      deletePostIdFetch(postId)
+    containerDiv.addEventListener('click', e => {
+      if (e.target === okDeleteBtn) {
+        console.log('bruhfuefu');
+        e.stopPropagation();
+        deletePostIdFetch(postId);
+      }
 
-    })
-
-
-
-    deleteModal.classList.remove('hidden');
+      if (e.target === e.currentTarget || e.target === cancelBtn) {
+        e.stopPropagation();
+        containerDiv.classList.add('hidden');
+      }
+    });
   });
 });
-//---------------------------------------------------------------/
-
-// router.delete(
-//   "/users/:id(\\d+)",
-//   requireAuth,
-//   asyncHandler(async (req, res, next) => {
-//     const userId = req.params.id * 1;
-
-//     const user = await db.User.findByPk(userId);
-
-//     if (req.session.auth.userId !== user.id) {
-//       const err = new Error("You are not authorized to delete this user.");
-//       err.status = 403;
-//       return next(err);
-//     }
-
-//     if (user) {
-//       user.activeState = false;
-//       await user.save();
-//       logoutUser(req, res);
-//       return res.json("success");
-//     } else {
-//       const err = new Error("User to delete was not found.");
-//       err.status = 404;
-//       next(err);
-//     }
-//   })
-// );
